Screenshot of app.
Here you can see the shadow above the toolbar, but I need it below the toolbar. I am using AppBarLayout.OnOffsetChangedListener to hide and show my title in the toolbar. If I removed that I get something like a shadow under the toolbar but the shadow above toolbar still exists.